modules/alarm/requirements.txt
PyYAML
paho-mqtt
python-dotenv
requests
redis
asyncpg
sqlalchemy[asyncio]
